use geotools 2.5.1 release instead of shaky snapshot

+++ trunk/gbif-providertool/src/test/java/org/gbif/provider/model/PointTest.java
@@ -9,17 +9,17 @@ public class PointTest {
 
 	@Test
 	public void rangeLimit(){
-		Point p1 = new Point(54f, 123f);
-		Point p2 = new Point(31.243f, -13.9883f);
+		Point p1 = new Point(54.0, 123.0);
+		Point p2 = new Point(31.243, -13.9883);
 		assertFalse(p1.equals(p2));
-		p1.setLatitude(31.243f);
-		p1.setLongitude(-13.9883f);
+		p1.setLatitude(31.243);
+		p1.setLongitude(-13.9883);
 		assertTrue(p1.equals(p2));
 		
 		p1.setLatitude(null);
 		boolean success=false;
 		try{
-			p1.setLatitude(121.243f);
+			p1.setLatitude(121.243);
 		}catch(IllegalArgumentException e){
 			success=true;
 		}
@@ -30,8 +30,8 @@ public class PointTest {
 		// test nulls
 		Point p = new Point();
 		Point pp = new Point(p);
-		Float lon = p.getLongitude();
-		Float lat = p.getLatitude();
+		Double lon = p.getLongitude();
+		Double lat = p.getLatitude();
 		pp = new Point(lat, lon);
 	}
 }
